(PC-24426)[PRO] refactor: move BookingsRecapTable components next to BookingsRecapTable
