+++ pro/src/screens/Bookings/BookingsRecapTable/Filters/FilterByOmniSearch/FilterByOmniSearch.module.scss
