+++ pro/src/screens/Bookings/BookingsRecapTable/Filters/__specs__/FilterByBookingStatus.spec.tsx
