+++ pro/src/screens/Bookings/BookingsRecapTable/Filters/__specs__/FilterByOmniSearch.spec.tsx
