+++ pro/src/screens/Bookings/BookingsRecapTable/NoFilteredBookings/__specs__/NoFilteredBookings.spec.tsx
